use config based rpc address
