+++ src/components/ConnectButton/index.tsx
@@ -2,9 +2,8 @@ import Onboard from 'bnc-onboard'
 import React from 'react'
 
 import Button from 'src/components/layout/Button'
-import { getNetworkId } from 'src/config'
+import { getNetworkId, getWeb3, setWeb3 } from 'src/config'
 import { ETHEREUM_NETWORK } from 'src/config/networks/network.d'
-import { getWeb3, setWeb3 } from 'src/logic/wallets/getWeb3'
 import { fetchProvider } from 'src/logic/wallets/store/actions'
 import transactionDataCheck from 'src/logic/wallets/transactionDataCheck'
 import { getSupportedWallets } from 'src/logic/wallets/utils/walletList'
