+++ src/config/index.ts
@@ -1,9 +1,13 @@
+import memoize from 'lodash.memoize'
+import Web3 from 'web3'
+import { provider as Provider } from 'web3-core'
+import { ContentHash } from 'web3-eth-ens'
+
 import networks from 'src/config/networks'
 import { EnvironmentSettings, ETHEREUM_NETWORK, NetworkSettings, SafeFeatures } from 'src/config/networks/network.d'
 import { checksumAddress } from 'src/utils/checksumAddress'
 import { APP_ENV, ETHERSCAN_API_KEY, GOOGLE_ANALYTICS_ID, INFURA_TOKEN, NETWORK, NODE_ENV } from 'src/utils/constants'
 import { ensureOnce } from 'src/utils/singleton'
-import memoize from 'lodash.memoize'
 
 export const getNetworkId = (): ETHEREUM_NETWORK => ETHEREUM_NETWORK[NETWORK]
 
@@ -172,3 +176,40 @@ export const getExplorerInfo = (hash: string): BlockScanInfo => {
 
   return blockScanInfo
 }
+
+// With some wallets from web3connect you have to use their provider instance only for signing
+// And our own one to fetch data
+export const web3ReadOnly = new Web3(
+  process.env.NODE_ENV !== 'test'
+    ? new Web3.providers.HttpProvider(getRpcServiceUrl())
+    : window.web3?.currentProvider || 'ws://localhost:8545',
+)
+
+let web3 = web3ReadOnly
+export const getWeb3 = (): Web3 => web3
+
+export const resetWeb3 = (): void => {
+  web3 = web3ReadOnly
+}
+
+export const setWeb3 = (provider: Provider): void => {
+  web3 = new Web3(provider)
+}
+
+export const getBalanceInEtherOf = async (safeAddress: string): Promise<string> => {
+  if (!web3) {
+    return '0'
+  }
+
+  const funds = await web3.eth.getBalance(safeAddress)
+
+  if (!funds) {
+    return '0'
+  }
+
+  return web3.utils.fromWei(funds, 'ether').toString()
+}
+
+export const getAddressFromENS = (name: string): Promise<string> => web3.eth.ens.getAddress(name)
+
+export const getContentFromENS = (name: string): Promise<ContentHash> => web3.eth.ens.getContenthash(name)
