+++ src/logic/safe/store/actions/fetchSafe.ts
@@ -5,7 +5,7 @@ import generateBatchRequests from 'src/logic/contracts/generateBatchRequests'
 import { getLocalSafe, getSafeName } from 'src/logic/safe/utils'
 import { enabledFeatures, safeNeedsUpdate } from 'src/logic/safe/utils/safeVersion'
 import { sameAddress } from 'src/logic/wallets/ethAddresses'
-import { getBalanceInEtherOf } from 'src/logic/wallets/getWeb3'
+import { getBalanceInEtherOf } from 'src/config'
 import addSafeOwner from 'src/logic/safe/store/actions/addSafeOwner'
 import removeSafeOwner from 'src/logic/safe/store/actions/removeSafeOwner'
 import updateSafe from 'src/logic/safe/store/actions/updateSafe'
