+++ src/logic/safe/transactions/multisend.ts
@@ -1,7 +1,7 @@
 import { Transaction } from '@gnosis.pm/safe-apps-sdk'
 import { AbiItem } from 'web3-utils'
 import { MultiSend } from 'src/types/contracts/MultiSend.d'
-import { getWeb3 } from 'src/logic/wallets/getWeb3'
+import { getWeb3 } from 'src/config'
 import { MULTI_SEND_ADDRESS } from 'src/logic/contracts/safeContracts'
 
 const multiSendAbi: AbiItem[] = [
