+++ src/logic/safe/utils/upgradeSafe.ts
@@ -7,7 +7,7 @@ import {
   getGnosisSafeInstanceAt,
 } from 'src/logic/contracts/safeContracts'
 import { DELEGATE_CALL } from 'src/logic/safe/transactions'
-import { getWeb3 } from 'src/logic/wallets/getWeb3'
+import { getWeb3 } from 'src/config'
 import { MultiSend } from 'src/types/contracts/MultiSend.d'
 
 interface MultiSendTx {
