+++ src/logic/wallets/ethTransactions.ts
@@ -1,7 +1,7 @@
 import axios from 'axios'
 import { BigNumber } from 'bignumber.js'
 
-import { getWeb3, web3ReadOnly } from 'src/logic/wallets/getWeb3'
+import { getWeb3, web3ReadOnly } from 'src/config'
 
 // const MAINNET_NETWORK = 1
 export const EMPTY_DATA = '0x'
