+++ src/logic/wallets/getWeb3.ts
@@ -1,12 +1,9 @@
 import Web3 from 'web3'
-import { provider as Provider } from 'web3-core'
-import { ContentHash } from 'web3-eth-ens'
 
-import { ETHEREUM_NETWORK } from 'src/config/networks/network.d'
 import { sameAddress } from './ethAddresses'
 import { EMPTY_DATA } from './ethTransactions'
 import { ProviderProps } from './store/model/provider'
-import { NODE_ENV, NETWORK, INFURA_TOKEN } from 'src/utils/constants'
+import { NODE_ENV } from 'src/utils/constants'
 
 export const WALLET_PROVIDER = {
   SAFE: 'SAFE',
@@ -24,39 +21,6 @@ export const WALLET_PROVIDER = {
   AUTHEREUM: 'AUTHEREUM',
   LEDGER: 'LEDGER',
   TREZOR: 'TREZOR',
-}
-
-export const getInfuraUrl = (network: ETHEREUM_NETWORK): string =>
-  `https://${ETHEREUM_NETWORK[network].toLowerCase()}.infura.io:443/v3/${INFURA_TOKEN}`
-
-export const getRPCUrl = (network: ETHEREUM_NETWORK): string => {
-  switch (network) {
-    case ETHEREUM_NETWORK.MAINNET:
-      return getInfuraUrl(network)
-    case ETHEREUM_NETWORK.RINKEBY:
-      return getInfuraUrl(network)
-    case ETHEREUM_NETWORK.ENERGY_WEB_CHAIN:
-      return 'https://rpc.energyweb.org'
-    case ETHEREUM_NETWORK.VOLTA:
-      return 'https://volta-rpc.energyweb.org'
-    default:
-      return ''
-  }
-}
-
-// With some wallets from web3connect you have to use their provider instance only for signing
-// And our own one to fetch data
-export const web3ReadOnly = new Web3(
-  process.env.NODE_ENV !== 'test'
-    ? new Web3.providers.HttpProvider(getRPCUrl(ETHEREUM_NETWORK[NETWORK] ?? ETHEREUM_NETWORK.RINKEBY))
-    : window.web3?.currentProvider || 'ws://localhost:8545',
-)
-
-let web3 = web3ReadOnly
-export const getWeb3 = (): Web3 => web3
-
-export const resetWeb3 = (): void => {
-  web3 = web3ReadOnly
 }
 
 export const getAccountFrom = async (web3Provider: Web3): Promise<string | null> => {
@@ -97,26 +61,4 @@ export const getProviderInfo = async (web3Instance: Web3, providerName = 'Wallet
     smartContractWallet,
     hardwareWallet,
   }
-}
-
-export const getAddressFromENS = (name: string): Promise<string> => web3.eth.ens.getAddress(name)
-
-export const getContentFromENS = (name: string): Promise<ContentHash> => web3.eth.ens.getContenthash(name)
-
-export const setWeb3 = (provider: Provider): void => {
-  web3 = new Web3(provider)
-}
-
-export const getBalanceInEtherOf = async (safeAddress: string): Promise<string> => {
-  if (!web3) {
-    return '0'
-  }
-
-  const funds = await web3.eth.getBalance(safeAddress)
-
-  if (!funds) {
-    return '0'
-  }
-
-  return web3.utils.fromWei(funds, 'ether').toString()
 }
