+++ src/logic/wallets/store/actions/fetchProvider.ts
@@ -2,11 +2,11 @@ import ReactGA from 'react-ga'
 
 import addProvider from './addProvider'
 
-import { getNetworkId } from 'src/config'
+import { getNetworkId, getWeb3 } from 'src/config'
 import { NOTIFICATIONS, enhanceSnackbarForAction } from 'src/logic/notifications'
 import enqueueSnackbar from 'src/logic/notifications/store/actions/enqueueSnackbar'
 import { ETHEREUM_NETWORK } from 'src/config/networks/network.d'
-import { getProviderInfo, getWeb3 } from 'src/logic/wallets/getWeb3'
+import { getProviderInfo } from 'src/logic/wallets/getWeb3'
 import { makeProvider } from 'src/logic/wallets/store/model/provider'
 import { updateStoredTransactionsStatus } from 'src/logic/safe/store/actions/transactions/utils/transactionHelpers'
 import { Dispatch } from 'redux'
