+++ src/routes/safe/components/Balances/SendModal/screens/ContractInteraction/utils/index.ts
@@ -3,7 +3,7 @@ import createDecorator from 'final-form-calculate'
 import { ContractSendMethod } from 'web3-eth-contract'
 
 import { AbiItemExtended } from 'src/logic/contractInteraction/sources/ABIService'
-import { getAddressFromENS, getWeb3 } from 'src/logic/wallets/getWeb3'
+import { getAddressFromENS, getWeb3 } from 'src/config'
 import { TransactionReviewType } from 'src/routes/safe/components/Balances/SendModal/screens/ContractInteraction/Review'
 import { isValidEnsName } from 'src/logic/wallets/ethAddresses'
