+++ src/routes/safe/components/Settings/ThresholdSettings/ChangeThreshold/index.tsx
@@ -19,7 +19,7 @@ import Row from 'src/components/layout/Row'
 import { getGnosisSafeInstanceAt } from 'src/logic/contracts/safeContracts'
 import { estimateTxGasCosts } from 'src/logic/safe/transactions/gasNew'
 import { formatAmount } from 'src/logic/tokens/utils/formatAmount'
-import { getWeb3 } from 'src/logic/wallets/getWeb3'
+import { getWeb3 } from 'src/config'
 
 const THRESHOLD_FIELD_NAME = 'threshold'
