+++ src/test/safe.dom.create.tsx
@@ -1,6 +1,4 @@
-// 
 import * as React from 'react'
-import { } from 'redux'
 import { render, fireEvent, act } from '@testing-library/react'
 import { Provider } from 'react-redux'
 import { ConnectedRouter } from 'connected-react-router'
@@ -8,7 +6,8 @@ import { sleep } from 'src/utils/timer'
 import { ADD_OWNER_BUTTON } from 'src/routes/open/components/SafeOwnersConfirmationsForm'
 import Open from 'src/routes/open/container/Open'
 import { aNewStore, history, } from 'src/store'
-import { getProviderInfo, getWeb3 } from 'src/logic/wallets/getWeb3'
+import { getWeb3 } from 'src/config'
+import { getProviderInfo} from 'src/logic/wallets/getWeb3'
 import addProvider from 'src/logic/wallets/store/actions/addProvider'
 import { makeProvider } from 'src/logic/wallets/store/model/provider'
 import { getGnosisSafeInstanceAt } from 'src/logic/contracts/safeContracts'
