+++ src/test/safe.dom.load.tsx
@@ -1,13 +1,12 @@
-// 
 import * as React from 'react'
-import { } from 'redux'
 import { Provider } from 'react-redux'
 import { render, fireEvent, act } from '@testing-library/react'
 import { ConnectedRouter } from 'connected-react-router'
 import Load from 'src/routes/load/container/Load'
 import { aNewStore, history, } from 'src/store'
 import { sleep } from 'src/utils/timer'
-import { getProviderInfo, getWeb3 } from 'src/logic/wallets/getWeb3'
+import { getWeb3 } from 'src/config'
+import { getProviderInfo } from 'src/logic/wallets/getWeb3'
 import addProvider from 'src/logic/wallets/store/actions/addProvider'
 import { makeProvider } from 'src/logic/wallets/store/model/provider'
 import { aMinedSafe } from './builder/safe.redux.builder'
