+++ src/test/utils/tokenMovements.ts
@@ -1,7 +1,7 @@
 //import contract from '@truffle/contract'
-import { getBalanceInEtherOf, getWeb3 } from 'src/logic/wallets/getWeb3'
+import { getBalanceInEtherOf, getWeb3 } from 'src/config'
 //import { ensureOnce } from 'src/utils/singleton'
-import { toNative } from 'src/logic/wallets/tokens'
+// import { toNative } from 'src/logic/wallets/tokens'
 //import TokenOMG from '../../../build/contracts/TokenOMG'
 //import TokenRDN from '../../../build/contracts/TokenRDN'
 //import Token6Decimals from '../../../build/contracts/Token6Decimals.json'
