+++ src/utils/checksumAddress.ts
@@ -1,4 +1,4 @@
-import { getWeb3 } from 'src/logic/wallets/getWeb3'
+import { getWeb3 } from 'src/config'
 
 export const checksumAddress = (address: string): string => {
   return getWeb3().utils.toChecksumAddress(address)
